+++ src/main/java/TubeSlice/tubeSlice/domain/oauth/OAuth2UserInfo.java
@@ -1,4 +1,4 @@
-package TubeSlice.tubeSlice.domain.user.dto.oauth;
+package TubeSlice.tubeSlice.domain.oauth;
 
 public interface OAuth2UserInfo {
     String getProvider();
