+++ src/main/java/TubeSlice/tubeSlice/domain/user/dto/response/UserResponseDto.java
@@ -17,7 +17,7 @@ public class UserResponseDto {
     private String code;
     private String message;
 
-    @JsonProperty(namespace = "result")
+    @JsonProperty("result")
     private LoginResponseDto loginResponseDto;
 
 }
